# workspace-server/docs/openapi/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "The per-tenant workspace-server HTTP API. Single source of truth for workspace/schedule/agent/secrets/files/memory CRUD. Hand-written clients (canvas, molecule-mcp-server, molecule-cli, molecule-ai-sdk) should be replaced by clients generated from this spec — see RFC #1706.",
        "title": "Molecule AI Workspace Server API",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.moleculesai.app",
    "basePath": "/",
    "paths": {
        "/admin/org/platform-agent/ensure": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "org"
                ],
                "summary": "Create or repair the org platform agent (concierge), core-only",
                "parameters": [
                    {
                        "description": "optional name/runtime/model/force",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/handlers.ensurePlatformAgentPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "422": {
                        "description": "RUNTIME_UNSUPPORTED or UNREGISTERED_MODEL_FOR_RUNTIME",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.ensurePlatformAgentPayload": {
            "type": "object",
            "properties": {
                "force": {
                    "description": "Force repairs (reinstall + re-provision) even a healthy 'online' concierge —\nthe explicit \"repair tool\" path. Default false keeps the call idempotent.",
                    "type": "boolean"
                },
                "model": {
                    "description": "Model, when set, is validated against the effective runtime and written as\nthe MODEL workspace_secret BEFORE the provision trigger fires, so the very\nfirst provision resolves this pick instead of racing it (core#3496 — the\nself-host onboarding scene's create path). Optional + additive: absent\nkeeps today's behavior (the platform default resolution at provision).",
                    "type": "string"
                },
                "name": {
                    "description": "Name is the concierge display name; defaults to defaultPlatformAgentName().",
                    "type": "string"
                },
                "runtime": {
                    "description": "Runtime is the concierge runtime; empty falls back to the platform default\n(conciergeDefaultRuntime), exactly like the CP-driven install. Used only on\na fresh insert — an existing root's runtime is preserved (runtime changes\ngo through the standard PATCH path). Must be a container-backed runtime:\nexternal-like/mock/unknown values are rejected 422 (platformRuntimeAllowed).",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token issued by Gitea (org-admin or persona PAT) or by the platform's signup/SSO flow.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OrgIdAuth": {
            "description": "Tenant routing header (UUID form). Alternative to X-Molecule-Org-Slug. At least one of OrgSlugAuth or OrgIdAuth must be sent alongside BearerAuth.",
            "type": "apiKey",
            "name": "X-Molecule-Org-Id",
            "in": "header"
        },
        "OrgSlugAuth": {
            "description": "Tenant routing header — required on every /workspaces/{id}/* request so the platform edge can route to the correct per-tenant workspace-server. Either X-Molecule-Org-Slug (human-readable, e.g. \"agents-team\") or X-Molecule-Org-Id (UUID) must be sent; slug is preferred for client code.",
            "type": "apiKey",
            "name": "X-Molecule-Org-Slug",
            "in": "header"
        }
    }
}